projectns: channel names will never have spaces
This fix uses the same approach as chanserv/register, i.e. asking for an
additional "trailing" parameter that we check is absent.

Apply the same check to the "project register" command to provide a
useful error message.

There are days I envy Limnoria's command argument decorators; this might
be one of them.