projectns: reorganize
The freenode/projectns module ceases to exist and is split into numerous
smaller modules that can be reloaded and maintained individually:

- freenode/projectns/main
- freenode/projectns/manage (REGISTER/DROP)
- freenode/projectns/set (SET OPENREG/REGINFO)
- freenode/projectns/hooks (for nickserv/info, chanserv/info,
                            chanserv/register)
- freenode/projectns/info
- freenode/projectns/help
- freenode/projectns/mark
- freenode/projectns/list
- freenode/projectns/channel
- freenode/projectns/contact
- freenode/projectns/cs_claim (ChanServ command for GCs to gain
                               channel access)

A projectserv is added to accomodate the various project-related
commands; it is only accessible to network staff.

Whether a channel is associated with a project, and the identity of the
project if so, are currently public in this implementation. This is
determined purely based on the namespace it falls in. Group contact
information is only available to users with project:auspex and the
individual user themselves.

Two privileges are added, project:auspex to view project-related
information and project:admin to alter it.

There are a few configuration options, all located within the
projectserv { } configuration block:

namespace_separators (string): a set of characters that indicate
a sub-channel within a namespace; e.g. "-." will allow #foo-bar
as well as #foo.bar

register_require_namespace (bool): if true, channels cannot be
registered without an existing project registration

register_require_namespace_exempt (string): if set, a wildcard mask
of channel names to exempt from the above check, e.g. "##*"

register_project_advice (string): a message to be displayed to users
trying to register a channel without an existing project registration