projectns: check names/namespaces for nonprintables on adding
Having non-printable characters (including color codes) in project names
or channel namespaces means they are hard to specify as parameters, and
they might creep in by accident (for example due to pasting from a
source that includes some invisible characters).

Checks are relaxed for attempts to delete projects/namespaces as they
might exist in the database for other reasons, and removing them should
still be possible in this case.