projectns: use contacts' visible/secondary flag
This will list the different types of contact in projectns/info output
as well as in nickserv/info output for staff and the user themselves.

The chanserv/info output will include a list of public contacts, if any;
for staff, private contacts will also be listed and the list will note
the "secondary" flag on any listed contacts.

The projectns/contact interface has also been extended to allow adding
contacts with the public or secondary flag. These flags can also be
modified on existing contacts.